course-project-group-49-main/backend/requirements.txt
firebase-admin
python-dotenv
flask
flask-restful
pyrebase4
jquery 
flask-cors